+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Protocol to query game server information (player count, version, etc.) and make ping tests.
 - Unlockable recipes
 - Localization support for prompt dialogs, diary sections, trade and group invitations.
+- Added Freezing Potion
 
 ### Changed